pixl_core/pyproject.toml
[project]
name = "core"
version = "0.0.1"
authors = [{ name = "PIXL core functionality" }]
description = ""
readme = "README.md"
requires-python = ">=3.9"
classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
    "fastapi==0.109.1",
    "token-bucket==0.3.0",
    "python-decouple==3.6",
    "python-slugify==8.0.1",
    "pika==1.3.1",
    "aio_pika==8.2.4",
    "environs==9.5.0",
    "requests==2.31.0",
    "jsonpickle==3.0.2",
    "sqlalchemy==2.0.24",
    "psycopg2-binary==2.9.9",
    "pandas==1.5.1",
    "pyarrow==14.0.1",
    "PyYAML==6.0.1",
]

[project.optional-dependencies]
test = [
    "pytest==7.4.2",
    "pytest-asyncio==0.21.1",
    "httpx==0.24.*",
    "pytest-pixl",
]
dev = ["mypy", "pre-commit", "ruff"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
markers = ["pika"]

[tool.ruff]
extend = "../ruff.toml"

[tool.ruff.extend-per-file-ignores]
"./tests/**" = ["D100"]

[tool.ruff.pep8-naming]
# Allow Pydantic's `@validator` decorator to trigger class method treatment.
classmethod-decorators = ["classmethod", "pydantic.field_validator"]
